Refactor channel selection into tabs
