+++ DemiCatPlugin/TemplatesWindow.cs
@@ -2,9 +2,11 @@ using System;
 using System.Net.Http;
 using System.Text;
 using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Numerics;
 using System.Threading.Tasks;
 using System.Linq;
+using System.Collections.Generic;
 using Dalamud.Bindings.ImGui;
 using DiscordHelper;
 
@@ -20,17 +22,40 @@ public class TemplatesWindow
     private EventView? _previewEvent;
     private TemplateType _selectedType;
     private string? _lastResult;
-
-    public string ChannelId { get; set; } = string.Empty;
+    private readonly List<ChannelDto> _channels = new();
+    private bool _channelsLoaded;
+    private bool _channelFetchFailed;
+    private int _channelIndex;
+    private string _channelId = string.Empty;
 
     public TemplatesWindow(Config config, HttpClient httpClient)
     {
         _config = config;
         _httpClient = httpClient;
+        _channelId = config.EventChannelId;
     }
 
     public void Draw()
     {
+        if (!_channelsLoaded)
+        {
+            _ = FetchChannels();
+        }
+        if (_channels.Count > 0)
+        {
+            var channelNames = _channels.Select(c => c.Name).ToArray();
+            if (ImGui.Combo("Channel", ref _channelIndex, channelNames, channelNames.Length))
+            {
+                _channelId = _channels[_channelIndex].Id;
+                _config.EventChannelId = _channelId;
+                SaveConfig();
+            }
+        }
+        else
+        {
+            ImGui.TextUnformatted(_channelFetchFailed ? "Failed to load channels" : "No channels available");
+        }
+
         _ = SignupPresetService.EnsureLoaded(_httpClient, _config);
         ImGui.BeginChild("TemplateList", new Vector2(150, 0), true);
         var typeNames = Enum.GetNames<TemplateType>();
@@ -109,9 +134,85 @@ public class TemplatesWindow
                 {
                     ImGui.TextUnformatted(_previewContent);
                 }
-            }
-            ImGui.End();
         }
+        ImGui.End();
+        }
+    }
+
+    private void SaveConfig()
+    {
+        PluginServices.Instance!.PluginInterface.SavePluginConfig(_config);
+    }
+
+    private async Task FetchChannels()
+    {
+        if (!ApiHelpers.ValidateApiBaseUrl(_config))
+        {
+            PluginServices.Instance!.Log.Warning("Cannot fetch channels: API base URL is not configured.");
+            _channelFetchFailed = true;
+            _channelsLoaded = true;
+            return;
+        }
+
+        try
+        {
+            var request = new HttpRequestMessage(HttpMethod.Get, $"{_config.ApiBaseUrl.TrimEnd('/')}/api/channels");
+            if (!string.IsNullOrEmpty(_config.AuthToken))
+            {
+                request.Headers.Add("X-Api-Key", _config.AuthToken);
+            }
+            var response = await _httpClient.SendAsync(request);
+            if (!response.IsSuccessStatusCode)
+            {
+                var responseBody = await response.Content.ReadAsStringAsync();
+                PluginServices.Instance!.Log.Warning($"Failed to fetch channels. Status: {response.StatusCode}. Response Body: {responseBody}");
+                _channelFetchFailed = true;
+                _channelsLoaded = true;
+                return;
+            }
+            var stream = await response.Content.ReadAsStreamAsync();
+            var dto = await JsonSerializer.DeserializeAsync<ChannelListDto>(stream) ?? new ChannelListDto();
+            ResolveChannelNames(dto.Event);
+            _ = PluginServices.Instance!.Framework.RunOnTick(() =>
+            {
+                _channels.Clear();
+                _channels.AddRange(dto.Event);
+                if (!string.IsNullOrEmpty(_channelId))
+                {
+                    _channelIndex = _channels.FindIndex(c => c.Id == _channelId);
+                    if (_channelIndex < 0) _channelIndex = 0;
+                }
+                if (_channels.Count > 0)
+                {
+                    _channelId = _channels[_channelIndex].Id;
+                }
+                _channelsLoaded = true;
+                _channelFetchFailed = false;
+            });
+        }
+        catch (Exception ex)
+        {
+            PluginServices.Instance!.Log.Error(ex, "Error fetching channels");
+            _channelFetchFailed = true;
+            _channelsLoaded = true;
+        }
+    }
+
+    private static void ResolveChannelNames(List<ChannelDto> channels)
+    {
+        foreach (var c in channels)
+        {
+            if (string.IsNullOrWhiteSpace(c.Name))
+            {
+                PluginServices.Instance!.Log.Warning($"Channel name missing for {c.Id}; using ID as fallback.");
+                c.Name = c.Id;
+            }
+        }
+    }
+
+    private class ChannelListDto
+    {
+        [JsonPropertyName("event")] public List<ChannelDto> Event { get; set; } = new();
     }
 
     private EmbedDto ToEmbedDto(Template tmpl)
@@ -146,9 +247,9 @@ public class TemplatesWindow
 
     private async Task PostTemplate(Template tmpl)
     {
-        if (!ApiHelpers.ValidateApiBaseUrl(_config) || string.IsNullOrWhiteSpace(ChannelId))
+        if (!ApiHelpers.ValidateApiBaseUrl(_config) || string.IsNullOrWhiteSpace(_channelId))
         {
-            if (string.IsNullOrWhiteSpace(ChannelId))
+            if (string.IsNullOrWhiteSpace(_channelId))
             {
                 _lastResult = "No channel selected";
             }
@@ -172,7 +273,7 @@ public class TemplatesWindow
 
                 var body = new
                 {
-                    channelId = ChannelId,
+                    channelId = _channelId,
                     title = tmpl.Title,
                     time = string.IsNullOrWhiteSpace(tmpl.Time)
                         ? DateTime.UtcNow.ToString("yyyy-MM-dd'T'HH:mm:ss.ffffff'Z'")
@@ -202,7 +303,7 @@ public class TemplatesWindow
             }
             else
             {
-                var body = new { channelId = ChannelId, content = tmpl.Content, useCharacterName = _config.UseCharacterName };
+                var body = new { channelId = _channelId, content = tmpl.Content, useCharacterName = _config.UseCharacterName };
                 var request = new HttpRequestMessage(HttpMethod.Post, $"{_config.ApiBaseUrl.TrimEnd('/')}/api/messages");
                 request.Content = new StringContent(JsonSerializer.Serialize(body), Encoding.UTF8, "application/json");
                 if (!string.IsNullOrEmpty(_config.AuthToken))
